requirements.txt
pandas
requests
python-dotenv
google-cloud-bigquery
pyarrow
dbt-core
dbt-bigquery
cryptography 
prefect